+++ src/app/static/src/tests/components/header/onlineSupportMenu.test.ts
@@ -138,6 +138,7 @@ describe("Online support menu", () => {
         expect(wrapper.find(ErrorReport).props("open")).toBe(false);
 
         const link = wrapper.findAll(".dropdown-item").at(1);
+
         expectTranslated(link, "Troubleshooting request", "Demande de dépannage", "Solicitação de solução de problemas", store as any);
 
         link.trigger("click");