+++ src/app/static/src/tests/genericChart/actions.test.ts
@@ -1,12 +1,14 @@
 import {mockAxios, mockFailure, mockGenericChartState, mockRootState, mockSuccess} from "../mocks";
 import {actions} from "../../app/store/genericChart/actions";
 import {GenericChartMutation} from "../../app/store/genericChart/mutations";
+import {freezer} from "../../app/utils";
 
 describe("genericChart actions", () => {
     beforeEach(() => {
         // stop apiService logging to console
         console.log = jest.fn();
         mockAxios.reset();
+        jest.clearAllMocks();
     });
 
     afterEach(() => {
@@ -16,13 +18,16 @@ describe("genericChart actions", () => {
     const rootState = mockRootState();
 
     it("gets generic chart metadata", async () => {
+        const mockResponse = mockSuccess("TEST METADATA");
         mockAxios.onGet("/meta/generic-chart")
-            .reply(200, mockSuccess("TEST METADATA"));
+            .reply(200, mockResponse);
         const commit = jest.fn();
+        const deepFreeze = jest.spyOn(freezer, "deepFreeze");
         await actions.getGenericChartMetadata({commit, rootState} as any);
         expect(commit.mock.calls.length).toEqual(1);
         expect(commit.mock.calls[0][0]["type"]).toBe("GenericChartMetadataFetched");
         expect(commit.mock.calls[0][0]["payload"]).toBe("TEST METADATA");
+        expect(deepFreeze).toHaveBeenCalledWith(mockResponse);
     });
 
     it("generic chart metadata action ignores errors",  async () => {
@@ -34,16 +39,19 @@ describe("genericChart actions", () => {
     });
 
     it("gets dataset", async () => {
+        const mockResponse = mockSuccess("TEST DATASET");
         mockAxios.onGet("/dataset1")
-            .reply(200, mockSuccess("TEST DATASET"));
+            .reply(200, mockResponse);
         const commit = jest.fn();
         const payload = {datasetId: "dataset1", url: "/dataset1"};
+        const deepFreeze = jest.spyOn(freezer, "deepFreeze");
         await actions.getDataset({commit, rootState} as any, payload);
         expect(commit.mock.calls.length).toEqual(2);
         expect(commit.mock.calls[0][0]["type"]).toBe(GenericChartMutation.SetError);
         expect(commit.mock.calls[0][0]["payload"]).toBeNull();
         expect(commit.mock.calls[1][0]["type"]).toBe(GenericChartMutation.SetDataset);
         expect(commit.mock.calls[1][0]["payload"]).toStrictEqual({dataset: "TEST DATASET", datasetId: "dataset1"});
+        expect(deepFreeze).toHaveBeenCalledWith(mockResponse);
     });
 
     it("sets error on get dataset", async () => {